Moved blog menu to component
