+++ lib/blog.ts
@@ -4,6 +4,30 @@ import matter from "gray-matter";
 
 const postsDirectory = join(process.cwd(), "_posts");
 
+export interface Post {
+  title : string;
+  description: string;
+  author: string;
+  date: string; 
+  photo: string;
+  topic: string;
+  featured: string;
+};
+
+export interface BlogProps {
+  posts: Post[];
+  topics: string[];
+  authors: IAuthor[];
+  featured: Post[];
+};
+
+export interface IAuthor {
+  author: string;
+  image: string;
+  posts: number;
+}
+
+
 export function getPostSlugs() {
   return fs
     .readdirSync(postsDirectory)
@@ -38,6 +62,28 @@ export function getPostBySlug(slug: string, fields: string[] = []) {
 
   return items;
 }
+
+export function getAuthors(posts) {
+  var result = [];
+
+  for(var i = 0; i < posts.length; i++) {
+    var found = false;
+    for(var j = 0; j < result.length; j++) {
+      if(result[j].author == posts[i].author) {
+        result[j].posts++;
+        found = true;
+        break;
+      }
+    }
+
+    if(!found) {
+      result.push({author: posts[i].author, image: posts[i].photo, posts: 1});
+    }
+  }
+
+  return result;
+}
+
 
 export function getAllPosts(fields: string[] = []) {
   const slugs = getPostSlugs();
